fix: 混沌工程审查 —— 修复生命周期竞态/数据边界/版本兼容三类隐性问题（P0×4/P1×4/P2×4）
生命周期竞态（P0/P1）：
- CyberNeonBallManager/BrowserOverlay/BrowserEngine：消灭 3 处孤儿
  CoroutineScope(Dispatchers.Main).launch，收敛为单例持有作用域；
  连点浮窗不再堆积无主协程抢 stateMutex
- PredictiveAttachmentPreprocessor：拷贝 Job 登记进缓存，cancel()/cancelAll()
  真正取消在途拷贝；进行中条目 putIfAbsent 登记，杜绝并发重复拷贝；
  文件名加 uri 哈希隔离 + 路径净化
- AgentChatViewModel：resume/retry/resumeCrashedTask 覆盖 currentJob 前先取消
- LogViewerScreen：导出改用 applicationContext，Activity 不再跨异步边界泄漏

数据一致性与边界值（P0/P1）：
- ToolOutputTruncator：head/tail 构造期收敛进 maxChars 预算
  （maxChars=200 合法输入下旧实现产物≈2×原文且省略计数为负）
- SlidingWindowCompressor：压缩边界回扩，保证 tool_call↔tool_result 配对，
  防止悬空 ToolResult 被持久化后每轮请求 400
- StreamingOpenAiClient：非流式解析容错（非 JSON 200 → LlmException.Http；
  usage 浮点 → intOrNull），不再绕过 ErrorClassifier 重试/降级体系
- RingTerminalBuffer.retainedBytes：Long 域取 min 再收窄，修复 2GiB 溢出为负

隐式依赖与版本兼容性（P0/P2）：
- AndroidManifest：补 VIBRATE 权限 —— 修复 WAITING_HUMAN 震动
  SecurityException 必现崩溃（详见报告用例 A）
- PermissionsScreen：Android 13+ 标准运行时请求 POST_NOTIFICATIONS，
  前台服务通知不再静默不可见
- app/build.gradle.kts：Shizuku 13.1.0 硬编码 → 版本目录 13.1.5 单一版本源

审查全文（含触发条件、崩溃路径、必现崩溃测试用例、遗留风险登记）：
docs/chaos-engineering-review.md

验证：
- core:logging / core:llm-adapter / core:tool-registry / core:agent-engine
  按 CI 同款 kotlinc 2.0.21 + classpath 本地编译通过
- 质量门（文件大小预算 / 反模式 / 括号平衡）全部通过

+++ platform/terminal/src/main/kotlin/com/apex/agent/platform/terminal/buffer/RingTerminalBuffer.kt
@@ -123,7 +123,10 @@ class RingTerminalBuffer(
 
     override val retainedBytes: Int
         get() = synchronized(lock) {
-            minOf(writePos.get().toInt(), capacityBytes)
+            // 混沌审查修复：writePos 是 Long 累计游标，长寿命会话跑高吞吐输出（yes/日志刷屏）
+            // 越过 2GiB 后旧实现 writePos.get().toInt() 回绕为负 → retainedBytes 变负数，
+            // 未来任何 retained/capacity 水位计算都得到负进度。先在 Long 域取 min 再收窄到 Int。
+            minOf(writePos.get(), capacityBytes.toLong()).toInt()
         }
 
     companion object {
